+++ runtime/test/src/lib.rs
@@ -753,7 +753,8 @@ impl_runtime_apis! {
 
     impl tp_consensus::TanssiAuthorityAssignmentApi<Block, NimbusId> for Runtime {
         /// Return the current authorities assigned to a given paraId
-        fn para_id_authorities(para_id: ParaId, parent_number: &sp_api::NumberFor<Block>) -> Option<Vec<NimbusId>> {
+        fn para_id_authorities(para_id: ParaId) -> Option<Vec<NimbusId>> {
+            let parent_number = System::block_number();
             let should_end_session = <Runtime as pallet_session::Config>::ShouldEndSession::should_end_session(parent_number + 1);
 
             let session_index = if should_end_session {
@@ -774,7 +775,8 @@ impl_runtime_apis! {
         }
 
         /// Return the paraId assigned to a given authority
-        fn check_para_id_assignment(authority: NimbusId, parent_number: &sp_api::NumberFor<Block>) -> Option<ParaId> {
+        fn check_para_id_assignment(authority: NimbusId) -> Option<ParaId> {
+            let parent_number = System::block_number();
             let should_end_session = <Runtime as pallet_session::Config>::ShouldEndSession::should_end_session(parent_number + 1);
 
             let session_index = if should_end_session {
